server/requirements.txt
Django==1.8.5
djangorestframework==3.2.4
Markdown==2.6.2
psycopg2==2.6.1
py==1.4.30
pytest==2.8.2
pytest-django==2.9.1
